requirements.txt
Flask==2.0.2
python-dotenv